Refactor import statements in fhir_agent.py and rag_chain.py
